Move User type into features/users
