+++ src/features/users/Admin/Admin.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 import AllUsers from "./AllUsers";
 import classes from './Admin.module.css'
 import UserDetails from "./UserDetails";
-import { User } from "../../../common/types/User";
+import { User } from "../types/User";
 import axios from "axios";
 import Card from "../../../common/components/Card/Card";
