+++ src/features/users/Admin/AllUsers.tsx
@@ -1,4 +1,4 @@
-import { User } from '../../../common/types/User';
+import { User } from '../types/User';
 import classes from './AllUsers.module.css'
 import UserItem from './UserItem';
